Fixes #23715 - Verify OS facts before submitting them
Before this PR, some roles would collect only a few facts from the
machine, but not 'os_major' 'os_minor' 'os_name', so these facts were
empty. The entire POST request would fail with a 500 in that case, as we
were submitting Operatingsystem.create! without any facts to create it.
